requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
transformers==4.41.0
torch==2.3.0
accelerate==0.30.0
sentencepiece==0.2.0
bitsandbytes==0.43.0
peft==0.11.0
Pillow